requirements.txt
absl-py==1.0.0
cachetools==5.0.0
captum==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
protobuf==3.19.4
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybaselines==0.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
PyYAML==6.0
renishawWiRE==0.1.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
